fix sort scoring for pv moves
for odd depths the pv of the previous depth would not be searched
first.
The cause of the error was storing the sort score in the Move
struct and comparing the current move being scored to the pv
move. If the pv move has a sort score from the previous depth it
would not be equal to a new move without a sort score and so pv
moves alternated having scores of 0 and 100.

The solution was to move the sort score to a separate array in the
MoveList struct. This has the added benefit of reducing the size
of the Move struct for storing in other locations without the need
for sorting.