Add search halting
Search function unravels when:
 - reaching max depth
 - reaching max nodes
 - stop flag set true

Main loop sets stop flag true when:
 - "stop" command input from user
 - allowed time has elapsed

The allowed time is calculated from a weighted
distribution that prioritises the middlegame.
This could be improved by adaptively increasring
or decreasing time in response to the position.

Iterative deepening will be required because the
current search will explore to max depth and
return garbage when halted.