Fix error reading promotions
The promoted piece in move notation was
capitlised for white and lowercase for black which
presented ambiguity when interfacing with a
program that did not follow this specification.

Since an incomming move string is compared to
each generated move string, mismatches in
capitlisation could cause the incomming move to
be considered illegal. In this case the engine
would play as the previous colour, outputing an
illegal move from the gui perspective.

All incomming and outgoing move strings are
converted to lowercase to remove ambiguity.